refactor: :recycle: create reusable pools table
