+++ packages/web/components/pages/asset-info-page/token-pools.tsx
@@ -1,0 +1,13 @@
+interface TokenPoolsProps {
+  denom: string;
+}
+
+export const TokenPools = (props: TokenPoolsProps) => {
+  const { denom } = props;
+
+  return (
+    <section>
+      <h5 className="mb-6">Pools</h5>
+    </section>
+  );
+};
